API/pyproject.toml
[project]
name = "api"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "diskcache>=5.6.3",
    "fastapi>=0.133.0",
    "litellm>=1.81.15",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.31.0",
    "scikit-learn>=1.8.0",
    "uvicorn[standard]>=0.41.0",
    "yfinance>=1.2.0",
]
